sprint 5

+++ src/main/java/com/paulmarten/availaball/service/FutsalFieldService.java
@@ -1,5 +1,6 @@
 package com.paulmarten.availaball.service;
 
+import javax.transaction.Transactional;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,8 +11,10 @@ import org.springframework.data.jpa.datatables.mapping.DataTablesOutput;
 import org.springframework.stereotype.Service;
 
 import com.paulmarten.availaball.model.Account;
+import com.paulmarten.availaball.model.DetailPrice;
 import com.paulmarten.availaball.model.FutsalField;
 import com.paulmarten.availaball.repository.AccountRepository;
+import com.paulmarten.availaball.repository.DetailPriceRepository;
 import com.paulmarten.availaball.repository.FutsalFieldRepository;
 
 /**
@@ -22,9 +25,12 @@ public class FutsalFieldService {
 
 	@Autowired
 	private FutsalFieldRepository futsalFieldRepository;
-	
+
 	@Autowired
 	private AccountRepository accountRepository;
+	
+	@Autowired
+	private DetailPriceRepository detailPriceRepository;
 	
 	public DataTablesOutput<FutsalField> findAllFutsalFieldAdmin(@Valid DataTablesInput input) {
 		return futsalFieldRepository.findAll(input);
@@ -39,12 +45,21 @@ public class FutsalFieldService {
 	}
 
 	public void deleteFutsalField(int id) {
-		futsalFieldRepository.delete(id);
+		FutsalField futsalFieldAccount = new FutsalField();
+		Account accountSave = new Account();
+		futsalFieldAccount = futsalFieldRepository.findOne(id);
+		accountSave = accountRepository.findOne(futsalFieldAccount.getAccount().getId());
+		int totalField = accountSave.getTotalField();
+		accountSave.setTotalField(totalField - 1);
+		if (accountRepository.save(accountSave) != null) {
+			futsalFieldRepository.delete(id);
+		}
+
 	}
 	
-	public void saveField(FutsalField futsalField){
+	public void saveField(FutsalField futsalField) {
 		java.util.Date today = new java.util.Date();
-		
+
 		FutsalField futsalFieldSave = new FutsalField();
 		futsalFieldSave.setFieldName(futsalField.getFieldName());
 		futsalFieldSave.setDetailLocation(futsalField.getDetailLocation());
@@ -56,13 +71,36 @@ public class FutsalFieldService {
 		futsalFieldSave.setAccount(futsalField.getAccount());
 		futsalFieldSave.setLocation(futsalField.getLocation());
 		futsalFieldSave.setLatestUpdate(new java.sql.Timestamp(today.getTime()));
-		
 		Account accountSaveFutsalField = new Account();
+		
+//		String[] days = futsalField.getDays().split(",");
+//		List<String> dayList = Arrays.asList(futsalField.getDays().split(","));
 		accountSaveFutsalField = accountRepository.findOne(futsalField.getAccount().getId());
-		if(futsalFieldRepository.save(futsalFieldSave)!= null){
+		if (futsalFieldRepository.save(futsalFieldSave) != null) {
+			FutsalField futsalFieldIdAfterSave = new FutsalField();
+			futsalFieldIdAfterSave = futsalFieldRepository.findTopByAccountOrderByIdFutsalFieldDesc(accountSaveFutsalField);
+//			System.out.println(futsalFieldIdAfterSave.getIdFutsalField());
 			int totalFutsalField = accountSaveFutsalField.getTotalField() + 1;
 			accountSaveFutsalField.setTotalField(totalFutsalField);
-			accountRepository.save(accountSaveFutsalField);
+			if(accountRepository.save(accountSaveFutsalField)!= null){
+				DetailPrice detailPrice = new DetailPrice();
+				String[]days = futsalField.getDays().split(",");
+				System.out.println(days.length);
+				for(int index = 0; index < days.length ; index ++){
+					System.out.println(days[index]);
+					detailPrice.setEndTime(futsalField.getClosingHours().toString());
+					detailPrice.setStartTime(futsalField.getOpeningHours().toString());
+					detailPrice.setPrice(futsalField.getPrice());
+					detailPrice.setDay(days[index]);
+					detailPrice.setFutsalField(futsalFieldIdAfterSave);
+					insert(detailPrice);
+				}
+			}
 		}
+	}
+	
+	@Transactional
+	public void insert(DetailPrice detailPrice) {
+		detailPriceRepository.save(detailPrice);
 	}
 }
